requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo
openai==1.55.3
langchain
langchain-community
langchain-openai
google-search-results
httpx==0.27.2
pytest
